NyayaAI-Indian-Legal-Assistance-main/NyayaAI-Indian-Legal-Assistance-main/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic==2.7.1

# RAG & Embeddings
llama-index==0.10.43
llama-index-vector-stores-faiss==0.1.2
llama-index-embeddings-huggingface==0.2.1
sentence-transformers
faiss-cpu==1.8.0
numpy<2.0.0

# Document processing
pypdf==4.2.0

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.9
